Build/Test Tools: Use Git when fetching the WordPress Importer for use in tests.
This switches to using Git in the local Docker environment install script to check out a copy of the WordPress Importer plugin for use in unit tests.

Previously, SVN was used and the commands were not correctly run within the Docker container. The container does not actually have SVN installed, and the script was only working when the machine running the command had SVN present.

Merges [51179] to the 5.1 branch.
Reviewed by desrosj.

Props czapla, alexstine, jnylen0, francina, desrosj.
See #52909.
See #62280.


git-svn-id: https://develop.svn.wordpress.org/branches/5.1@59615 602fd350-edb4-49c9-b593-d223f7449a82