Build/Test Tools: Update `@playwright/test`.
This updates `@playwright/test` to the latest version, currently `1.49.1`.

In older branches using Playwright, the E2E and Performance workflows have recently started failing. This is due to changes in the GitHub Actions runner images. Updating Playwright ensures more modern dependency trees are used when installing browsers for testing and fixes the issue.

Merges [59682] to the 6.7 branch.

Props swissspidy.
See #62843.

git-svn-id: https://develop.svn.wordpress.org/branches/6.7@59683 602fd350-edb4-49c9-b593-d223f7449a82