HTML API: Ensure that full processor can seek to earlier bookmarks.
When the HTML Processor seeks to an earlier place, it returns the the beginning of the document and proceeds forward until it reaches the appropriate location. This requires resetting internal state so that the processor can correctly proceed from the beginning of the document.

The seeking reset logic was not adapted to account for the full processor (i.e. when created via WP_HTML_Processor::create_full_parser()). This change updates the seek logic to account for the full and fragment parsers as well as other state that has been introduced in the interim and should be reset.

Reviewed by westonruter, jorbin.
Merges [59391] to the 6.7 branch.

Props jonsurrell, dmsnell, westonruter, mi5t4n, bernhard-reiter.
Fixes #62290.


git-svn-id: https://develop.svn.wordpress.org/branches/6.7@59721 602fd350-edb4-49c9-b593-d223f7449a82