REST API: Protect against fatal error for post types without format support.
Ignore the `format` parameter introduced in WordPress 6.7 for post types that do not support post formats. This protects against a fatal error being thrown in later version of PHP or a warning in earlier versions of PHP.

Follow up to r59115.

Reviewed by jorbin.
Merges [59544] to the 6.7 branch.

Props dd32, sergeybiryukov, yogeshbhutkar.
Fixes #62646.
See #62014.



git-svn-id: https://develop.svn.wordpress.org/branches/6.7@59744 602fd350-edb4-49c9-b593-d223f7449a82