Docs: Fix types for `wp_cache_set_last_changed` filter.
Corrects the type for the `$time` and `$previous_time` parameters in the filter to indicate the times are expressed as strings. The values are generated from a call to `microtime()` which returns the time as a string in the form `msec sec`.

Reviewed by audrasjb.
Merges [60128] to the 6.8 branch.

Props tillkruess, westonruter.
Fixes #63194.



git-svn-id: https://develop.svn.wordpress.org/branches/6.8@60133 602fd350-edb4-49c9-b593-d223f7449a82