Bundled Themes: Bump `Tested up to` for Twenty Twenty-Three & Twenty Twenty-Four.
An update for these two default themes will not be released in coordination with WordPress 6.8 due because there were no changes to the source code during the release cycle. However, they should be considered tested with 6.8, and this ensures that the next release properly indicates that.

Reviewed by joemcgill.
Merges [60158] to the 6.8 branch.

Props shailu25, sabernhardt, desrosj, poena, audrasjb.
See #62990.

git-svn-id: https://develop.svn.wordpress.org/branches/6.8@60161 602fd350-edb4-49c9-b593-d223f7449a82