Upgrade/Install: Pass stored credentials to WP_Filesystem() in WP_Upgrader.
When using a non-direct filesystem, the call in WP_Upgrader::maintenance_mode() did not include the required credentials, leading to a fatal error as the connection was not initialized properly.

This commit attempts to use the stored credentials if available, and triggers a notice otherwise.

Follow-up to [56341], [58128].

Reviewed by jorbin.
Merges [60107] to the 6.8 branch.

Props SirLouen, hideishi, dd32, tusharaddweb, takuword, SergeyBiryukov.
Fixes #62718.


git-svn-id: https://develop.svn.wordpress.org/branches/6.8@60184 602fd350-edb4-49c9-b593-d223f7449a82