Build/Test Tools: Test version in additional config files.
Introduces tests to ensure the version numbers in `package-lock.json` (two instances) and `composer.json` match the WordPress version specified in `version.php`.

In pull requests, the `package-lock.json` file is updated automatically to match the version in `package.json`. The new test is intended to ensure the version numbers are correct in production branches.

Reviewed by jorbin.
Merges [60219], [60221] to the 6.8 branch.

Fixes #63390.


git-svn-id: https://develop.svn.wordpress.org/branches/6.8@60223 602fd350-edb4-49c9-b593-d223f7449a82