Coding Standards: Simplify a conditional in `get_url_in_content()`.
As there's an `is_string()` check already, `! empty( $href )` can be simplified to a string comparison, as the other variable types that are checked in `empty()` won't appear.

`empty()` also returns `false` for the string `"0"` which would however be a valid (relative) URL and thus should be detectable by the function.

Follow-up to [60665].

Props TobiasBg.
Fixes #63694.

git-svn-id: https://develop.svn.wordpress.org/trunk@60726 602fd350-edb4-49c9-b593-d223f7449a82