Build/Test Tools: Configure logging for PHPUnit and multisite.
This ensures a log file is created after running the PHPUnit test suite as a multisite.

In practice, this file is used by the Host Test Runner to collect a list of failing tests to submit to wordpress.org.

See https://github.com/WordPress/phpunit-test-runner/issues/267.

Props rishabhwp, nimeshatxecurify.
Fixes #63884.

git-svn-id: https://develop.svn.wordpress.org/trunk@60781 602fd350-edb4-49c9-b593-d223f7449a82