REST API: Rename post link relation in media endpoint to wp:attached-to
By adding the https://api.w.org/ prefix and using a more descriptive name for the link relation, this change creates greater clarity for this part of the API. Note that the renaming of this property is safe, as the API change has not yet made it into a final release.

Follow-up to [60893].

Props andrewserong, mukesh27, ramonopoly, timothyblynjacobs.
Fixes #64034.



git-svn-id: https://develop.svn.wordpress.org/trunk@60916 602fd350-edb4-49c9-b593-d223f7449a82